Crate simd

Functionality for working with SIMD registers and instructions.

Source, crates.io

Modules

maths
Mathematical operations.

Structs

Tails
A type that repesents a &[U] converted from a &[T], where the conversion may be forced to leave leading/trailing elements (due to alignment/size mismatches).

TailsMut
A type that repesents a &mut [U] converted from a &mut [T], where the conversion may be forced to leave leading/trailing elements (due to alignment/size mismatches).

Traits

Bitcast
Types between which a direct bitcast (transmute) is safe, e.g. reinterpreting the 4 bytes of a u32 as an i32 is fine.

Convert

DoubleVector
SIMD vectors which can be merged with another of the same type to create one of double the length with the same elements.

HalfVector
SIMD vectors which can be separated into two SIMD vectors of half the size with the same elements.

Vector
A SIMD vector, storing precisely count() elements of (primitive) type Item sequentially in memory.

Functions

rsqrt
Return a platform-specific approximation to the reciprocal-square root of x.

sqrt
Return a platform-specific approximation to the square root of x.