Struct simd::aarch64::neon::i64x2

#[repr(simd)]

pub struct i64x2(_, _);

A SIMD vector of 2 i64s.

Methods

impl i64x2

const fn new(x0: i64, x1: i64) -> i64x2

Create a new instance.

const fn splat(x: i64) -> i64x2

Create a new instance where every lane has value x.

fn eq(self, other: Self) -> bool64ix2

Compare for equality.

fn ne(self, other: Self) -> bool64ix2

Compare for equality.

fn lt(self, other: Self) -> bool64ix2

Compare for equality.

fn le(self, other: Self) -> bool64ix2

Compare for equality.

fn gt(self, other: Self) -> bool64ix2

Compare for equality.

fn ge(self, other: Self) -> bool64ix2

Compare for equality.

fn extract(self, idx: u32) -> i64

Extract the value of the idxth lane of self.

Panics

extract will panic if idx is out of bounds.

fn replace(self, idx: u32, elem: i64) -> Self

Return a new vector where the idxth lane is replaced by elem.

Panics

replace will panic if idx is out of bounds.

fn load(array: &[i64], idx: usize) -> Self

Load a new value from the idxth position of array.

This is equivalent to the following, but is possibly more efficient:

Self::new(array[idx], array[idx + 1], ...)

Panics

load will panic if idx is out of bounds in array, or if array[idx..] is too short.

fn store(self, array: &mut [i64], idx: usize)

Store the elements of self to array, starting at the idxth position.

This is equivalent to the following, but is possibly more efficient:

array[i] = self.extract(0);
array[i + 1] = self.extract(1);
// ...

Panics

store will panic if idx is out of bounds in array, or if array[idx...] is too short.

impl i64x2

fn to_u64(self) -> u64x2

Convert each lane to an unsigned integer.

fn to_f64(self) -> f64x2

Convert each lane to a 64-bit float.