Struct simd::bool32fx4

#[repr(simd)]

pub struct bool32fx4(_, _, _, _);

A SIMD boolean vector for length-4 vectors of 32-bit floats.

Methods

impl bool32fx4

pub fn to_repr(self) -> i32x4

Convert to integer representation.

pub fn from_repr(x: i32x4) -> Self

Convert from integer representation.

pub fn new(x0: bool, x1: bool, x2: bool, x3: bool) -> bool32fx4

Create a new instance.

pub fn splat(x: bool) -> bool32fx4

Create a new instance where every lane has value x.

pub fn extract(self, idx: u32) -> bool

Extract the value of the idxth lane of self.

Panics

extract will panic if idx is out of bounds.

pub fn replace(self, idx: u32, elem: bool) -> Self

Return a new vector where the idxth lane is replaced by elem.

Panics

replace will panic if idx is out of bounds.

pub fn select<T: Simd<Bool = bool32fx4>>(self, then: T, else_: T) -> T

Select between elements of then and else_, based on the corresponding element of self.

This is equivalent to the following, but is possibly more efficient:

T::new(if self.extract(0) { then.extract(0) } else { else_.extract(0) },
       if self.extract(1) { then.extract(1) } else { else_.extract(1) },
       ...)

pub fn all(self) -> bool

Check if every element of self is true.

This is equivalent to the following, but is possibly more efficient:

self.extract(0) && self.extract(1) && ...

pub fn any(self) -> bool

Check if any element of self is true.

This is equivalent to the following, but is possibly more efficient:

self.extract(0) || self.extract(1) || ...

pub fn to_i(self) -> bool32ix4

Convert self to a boolean vector for interacting with integer vectors.