Struct simd::f32x4

pub struct f32x4(_, _, _, _);

A SIMD vector of 4 f32s.

Methods

impl f32x4

const fn new(x0: f32, x1: f32, x2: f32, x3: f32) -> f32x4

Create a new instance.

const fn splat(x: f32) -> f32x4

Create a new instance where every lane has value x.

fn eq(self, other: Self) -> bool32fx4

Compare for equality.

fn ne(self, other: Self) -> bool32fx4

Compare for equality.

fn lt(self, other: Self) -> bool32fx4

Compare for equality.

fn le(self, other: Self) -> bool32fx4

Compare for equality.

fn gt(self, other: Self) -> bool32fx4

Compare for equality.

fn ge(self, other: Self) -> bool32fx4

Compare for equality.

fn extract(self, idx: u32) -> f32

Extract the value of the idxth lane of self.

Panics

extract will panic if idx is out of bounds.

fn replace(self, idx: u32, elem: f32) -> Self

Return a new vector where the idxth lane is replaced by elem.

Panics

replace will panic if idx is out of bounds.

fn load(array: &[f32], idx: usize) -> Self

Load a new value from the idxth position of array.

This is equivalent to the following, but is possibly more efficient:

Self::new(array[idx], array[idx + 1], ...)

Panics

load will panic if idx is out of bounds in array, or if array[idx..] is too short.

fn store(self, array: &mut [f32], idx: usize)

Store the elements of self to array, starting at the idxth position.

This is equivalent to the following, but is possibly more efficient:

array[i] = self.extract(0);
array[i + 1] = self.extract(1);
// ...

Panics

store will panic if idx is out of bounds in array, or if array[idx...] is too short.

impl f32x4

fn sqrt(self) -> Self

Compute the square root of each lane.

fn approx_rsqrt(self) -> Self

Compute an approximation to the reciprocal of the square root of self, that is, f32::splat(1.0) / self.sqrt().

The accuracy of this approximation is platform dependent.

fn approx_reciprocal(self) -> Self

Compute an approximation to the reciprocal of self, that is, f32::splat(1.0) / self.

The accuracy of this approximation is platform dependent.

fn max(self, other: Self) -> Self

Compute the lane-wise maximum of self and other.

This is equivalent to the following, but is possibly more efficient:

f32x4::new(self.extract(0).max(other.extract(0)),
           self.extract(1).max(other.extract(1)),
           ...)

fn min(self, other: Self) -> Self

Compute the lane-wise minimum of self and other.

This is equivalent to the following, but is possibly more efficient:

f32x4::new(self.extract(0).min(other.extract(0)),
           self.extract(1).min(other.extract(1)),
           ...)

fn to_i32(self) -> i32x4

Convert each lane to a signed integer.

fn to_u32(self) -> u32x4

Convert each lane to an unsigned integer.