Struct simd::i32x4

pub struct i32x4(_, _, _, _);

A SIMD vector of 4 i32s.

Methods

impl i32x4

const fn new(x0: i32, x1: i32, x2: i32, x3: i32) -> i32x4

Create a new instance.

const fn splat(x: i32) -> i32x4

Create a new instance where every lane has value x.

fn eq(self, other: Self) -> bool32ix4

Compare for equality.

fn ne(self, other: Self) -> bool32ix4

Compare for equality.

fn lt(self, other: Self) -> bool32ix4

Compare for equality.

fn le(self, other: Self) -> bool32ix4

Compare for equality.

fn gt(self, other: Self) -> bool32ix4

Compare for equality.

fn ge(self, other: Self) -> bool32ix4

Compare for equality.

fn extract(self, idx: u32) -> i32

Extract the value of the idxth lane of self.

Panics

extract will panic if idx is out of bounds.

fn replace(self, idx: u32, elem: i32) -> Self

Return a new vector where the idxth lane is replaced by elem.

Panics

replace will panic if idx is out of bounds.

fn load(array: &[i32], idx: usize) -> Self

Load a new value from the idxth position of array.

This is equivalent to the following, but is possibly more efficient:

Self::new(array[idx], array[idx + 1], ...)

Panics

load will panic if idx is out of bounds in array, or if array[idx..] is too short.

fn store(self, array: &mut [i32], idx: usize)

Store the elements of self to array, starting at the idxth position.

This is equivalent to the following, but is possibly more efficient:

array[i] = self.extract(0);
array[i + 1] = self.extract(1);
// ...

Panics

store will panic if idx is out of bounds in array, or if array[idx...] is too short.

impl i32x4

fn to_u32(self) -> u32x4

Convert each lane to an unsigned integer.

fn to_f32(self) -> f32x4

Convert each lane to a 32-bit float.