Struct simd::i8x16

pub struct i8x16(_, _, _, _, _, _, _, _, _, _, _, _, _, _, _, _);

A SIMD vector of 16 i8s.

Methods

impl i8x16

const fn new(x0: i8, x1: i8, x2: i8, x3: i8, x4: i8, x5: i8, x6: i8, x7: i8, x8: i8, x9: i8, x10: i8, x11: i8, x12: i8, x13: i8, x14: i8, x15: i8) -> i8x16

Create a new instance.

const fn splat(x: i8) -> i8x16

Create a new instance where every lane has value x.

fn eq(self, other: Self) -> bool8ix16

Compare for equality.

fn ne(self, other: Self) -> bool8ix16

Compare for equality.

fn lt(self, other: Self) -> bool8ix16

Compare for equality.

fn le(self, other: Self) -> bool8ix16

Compare for equality.

fn gt(self, other: Self) -> bool8ix16

Compare for equality.

fn ge(self, other: Self) -> bool8ix16

Compare for equality.

fn extract(self, idx: u32) -> i8

Extract the value of the idxth lane of self.

Panics

extract will panic if idx is out of bounds.

fn replace(self, idx: u32, elem: i8) -> Self

Return a new vector where the idxth lane is replaced by elem.

Panics

replace will panic if idx is out of bounds.

fn load(array: &[i8], idx: usize) -> Self

Load a new value from the idxth position of array.

This is equivalent to the following, but is possibly more efficient:

Self::new(array[idx], array[idx + 1], ...)

Panics

load will panic if idx is out of bounds in array, or if array[idx..] is too short.

fn store(self, array: &mut [i8], idx: usize)

Store the elements of self to array, starting at the idxth position.

This is equivalent to the following, but is possibly more efficient:

array[i] = self.extract(0);
array[i + 1] = self.extract(1);
// ...

Panics

store will panic if idx is out of bounds in array, or if array[idx...] is too short.

impl i8x16

fn to_u8(self) -> u8x16

Convert each lane to an unsigned integer.