Trait simd::Vector

pub unsafe trait Vector: Sized + Copy {
    type Item;
    fn count(_: Option<Self>) -> usize { ... }
}

A SIMD vector, storing precisely count() elements of (primitive) type Item sequentially in memory.

This trait is unsafe because it is used inside unsafe code relying on the correctness of Item and count().

Associated Types

type Item

The type that this vector contains.

Provided Methods

fn count(_: Option<Self>) -> usize

Return the number of items in self.