Struct simd::x86::avx::bool64fx4

#[repr(simd)]

pub struct bool64fx4(_, _, _, _);

Methods

impl bool64fx4

fn to_repr(self) -> i64x4

Convert to integer representation.

fn from_repr(x: i64x4) -> Self

Convert from integer representation.

fn new(x0: bool, x1: bool, x2: bool, x3: bool) -> bool64fx4

Create a new instance.

fn splat(x: bool) -> bool64fx4

Create a new instance where every lane has value x.

fn extract(self, idx: u32) -> bool

Extract the value of the idxth lane of self.

Panics

extract will panic if idx is out of bounds.

fn replace(self, idx: u32, elem: bool) -> Self

Return a new vector where the idxth lane is replaced by elem.

Panics

replace will panic if idx is out of bounds.

fn select<T: Simd<Bool=bool64fx4>>(self, then: T, else_: T) -> T

Select between elements of then and else_, based on the corresponding element of self.

This is equivalent to the following, but is possibly more efficient:

T::new(if self.extract(0) { then.extract(0) } else { else_.extract(0) },
       if self.extract(1) { then.extract(1) } else { else_.extract(1) },
       ...)

fn all(self) -> bool

Check if every element of self is true.

This is equivalent to the following, but is possibly more efficient:

self.extract(0) && self.extract(1) && ...

fn any(self) -> bool

Check if any element of self is true.

This is equivalent to the following, but is possibly more efficient:

self.extract(0) || self.extract(1) || ...

fn to_i(self) -> bool64ix4

Convert self to a boolean vector for interacting with integer vectors.