Struct simd::x86::avx::f32x8

#[repr(simd)]

pub struct f32x8(_, _, _, _, _, _, _, _);

Methods

impl f32x8

const fn new(x0: f32,
x1: f32,
x2: f32,
x3: f32,
x4: f32,
x5: f32,
x6: f32,
x7: f32)
-> f32x8

Create a new instance.

const fn splat(x: f32) -> f32x8

Create a new instance where every lane has value x.

fn eq(self, other: Self) -> bool32fx8

Compare for equality.

fn ne(self, other: Self) -> bool32fx8

Compare for equality.

fn lt(self, other: Self) -> bool32fx8

Compare for equality.

fn le(self, other: Self) -> bool32fx8

Compare for equality.

fn gt(self, other: Self) -> bool32fx8

Compare for equality.

fn ge(self, other: Self) -> bool32fx8

Compare for equality.

fn extract(self, idx: u32) -> f32

Extract the value of the idxth lane of self.

Panics

extract will panic if idx is out of bounds.

fn replace(self, idx: u32, elem: f32) -> Self

Return a new vector where the idxth lane is replaced by elem.

Panics

replace will panic if idx is out of bounds.

fn load(array: &[f32], idx: usize) -> Self

Load a new value from the idxth position of array.

This is equivalent to the following, but is possibly more efficient:

Self::new(array[idx], array[idx + 1], ...)

Panics

load will panic if idx is out of bounds in array, or if array[idx..] is too short.

fn store(self, array: &mut [f32], idx: usize)

Store the elements of self to array, starting at the idxth position.

This is equivalent to the following, but is possibly more efficient:

array[i] = self.extract(0);
array[i + 1] = self.extract(1);
// ...

Panics

store will panic if idx is out of bounds in array, or if array[idx...] is too short.