Struct simd::x86::sse2::u64x2

pub struct u64x2(_, _);

A SIMD vector of 2 u64s.

Methods

impl u64x2

const fn new(x0: u64, x1: u64) -> u64x2

Create a new instance.

const fn splat(x: u64) -> u64x2

Create a new instance where every lane has value x.

fn eq(self, other: Self) -> bool64ix2

Compare for equality.

fn ne(self, other: Self) -> bool64ix2

Compare for equality.

fn lt(self, other: Self) -> bool64ix2

Compare for equality.

fn le(self, other: Self) -> bool64ix2

Compare for equality.

fn gt(self, other: Self) -> bool64ix2

Compare for equality.

fn ge(self, other: Self) -> bool64ix2

Compare for equality.

fn extract(self, idx: u32) -> u64

Extract the value of the idxth lane of self.

Panics

extract will panic if idx is out of bounds.

fn replace(self, idx: u32, elem: u64) -> Self

Return a new vector where the idxth lane is replaced by elem.

Panics

replace will panic if idx is out of bounds.

fn load(array: &[u64], idx: usize) -> Self

Load a new value from the idxth position of array.

This is equivalent to the following, but is possibly more efficient:

Self::new(array[idx], array[idx + 1], ...)

Panics

load will panic if idx is out of bounds in array, or if array[idx..] is too short.

fn store(self, array: &mut [u64], idx: usize)

Store the elements of self to array, starting at the idxth position.

This is equivalent to the following, but is possibly more efficient:

array[i] = self.extract(0);
array[i + 1] = self.extract(1);
// ...

Panics

store will panic if idx is out of bounds in array, or if array[idx...] is too short.

impl u64x2

fn to_i64(self) -> i64x2

Convert each lane to a signed integer.

fn to_f64(self) -> f64x2

Convert each lane to a 64-bit float.